+++ src/wp-rest-loop/Wp-rest-loop.js
@@ -13,8 +13,8 @@ import Vue from "vue";
 export default {
   install: (Vue, options) => {
     /**
-    * Loop of Posts
-    * @Event Call GetPosts() function form file ./GetPosts.js
+    * Get site endpoint
+    * @Event Call get_endpoint() function from file ./Config.js
     * @return string of endpoint of site
     */
     Vue.prototype.$get_endpoint = () => {
@@ -22,21 +22,38 @@ export default {
     },
 
     /**
+    * Wp Request Api
+    * @Event Call Reques() function from file ./GetPosts.js
+    * @return object JSON
+    */
+    Vue.prototype.$wp_rest_request = (args, url) => {
+      return Index.methods.wp_rest_request(args, url);
+    },
+
+    /**
     * Loop of Posts
     * @Event Call GetPosts() function form file ./GetPosts.js
     * @return array of posts[]
     */
-    Vue.prototype.$get_posts = (args, endPoint) => {
-      return Index.methods.get_posts(args, endPoint);
+    Vue.prototype.$get_posts = (args) => {
+      return Index.methods.get_posts(args);
     },
 
+    /**
+    * Chek if have Thumbnail
+    * @Event Call GetThumbnail() function form file ./GetThumbnail.js
+    * @return The link src of the featured Thumbnail
+    */
+    Vue.prototype.$has_thumbnail = (post, size, get) => {
+      return Index.methods.has_thumbnail(post);
+    }
     /**
     * Recive the Thumbnail
     * @Event Call GetThumbnail() function form file ./GetThumbnail.js
     * @return The link src of the featured Thumbnail
     */
-    Vue.prototype.$get_thumbnail = (post, size) => {
-      return Index.methods.get_thumbnail(post, size);
+    Vue.prototype.$get_thumbnail = (post, size, get) => {
+      return Index.methods.get_thumbnail(post, size, get);
     }
 
     /**