+++ src/wp-rest-loop/index.js
@@ -3,9 +3,10 @@
 * @return all function of aplication
 */
 import GetPosts     from "./GetPosts";
-import GetThumbnail from "./GetThumbnail";
+import {GetThumbnail, Hasthumbnail} from "./GetThumbnail";
 import GetMenu      from "./GetMenu";
 import Config       from "./Config";
+import Request     from "./Request";
 
 export default {
 
@@ -13,10 +14,11 @@ export default {
   * @return The siteUrl + Json endPoint of site
   */
   methods:{
-    get_endpoint:   Config.get_EndPoint(),
-    get_posts:      GetPosts,
-    get_thumbnail:  GetThumbnail,
-    get_menu_itens: GetMenu,
-
+    get_endpoint:     Config.get_EndPoint,
+    wp_rest_request:  Request,
+    get_posts:        GetPosts,
+    has_thumbnail:    Hasthumbnail,
+    get_thumbnail:    GetThumbnail,
+    get_menu_itens:   GetMenu,
   }
 };