+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.3] - 2019-03-9
+
+### Added
+- /src/wp-rest-loop/Request
+- New Funcion $wp_rest_request(args, url) / Request(args, url) / wp_rest_request(args, url)
+- New Funcion has_thumbnail(post) / has_thumbnail(post)
+
+### Changed
+- All requests functions are now asynchronous
+  - See the wiki to know the new methods
+
+
 ## [1.0.2] - 2019-03-1
 
 ### Added